services/ingest_func/requirements.txt
azure-functions
azure-storage-blob>=12.19.0
azure-ai-documentintelligence>=1.0.0
azure-identity>=1.15.0
pydantic>=2.0.0
